backend/requirements.txt
google-genai
python-dotenv
fastapi==0.115.5
uvicorn==0.32.1
pydantic>=2.0.0
pinecone>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
slowapi>=0.1.9
